rdd-lab2

+++ src/main/scala/rdd/model/Product.scala
@@ -3,6 +3,6 @@ package rdd.model
 case class Product(
                     id: Int,
                     name: String,
-                    price: Int,
+                    price: Double,
                     numberOfProducts: Int
                   )
